Change rating type

+++ internal/yts/yts_modle.go
@@ -22,7 +22,7 @@ type Movie struct {
 	TitleLong               string    `json:"title_long"`
 	Slug                    string    `json:"slug"`
 	Year                    int       `json:"year"`
-	Rating                  float32   `json:"rating"`
+	Rating                  int       `json:"rating"`
 	Runtime                 int       `json:"runtime"`
 	Genres                  []string  `json:"genres"`
 	Summary                 string    `json:"summary"`
